#43 Simplify code
