+++ pygeppetto/services/data_source_service.py
@@ -4,18 +4,9 @@ from pygeppetto.model.exceptions import GeppettoInitializationException
 from pygeppetto.model.model_access import GeppettoModelAccess
 from pygeppetto.model.utils.datasource import query_check
 from pygeppetto.visitors.data_source_visitors import ExecuteQueryVisitor
+from .utils import set_custom_query_result_hash, unset_custom_query_result_hash
 
-# Use to distinguish QueryResult by ID in a set
-def modified_hash(self):
-    return hash(self.values[self._id_pos])
-def modified_eq(self, other):
-    return self.values[self._id_pos] == other.values[self._id_pos]
-
-# Restore default set identification
-def regular_hash(self):
-    return super(AQueryResult, self).__hash__()
-def regular_eq(self, other):
-    return super(AQueryResult, self).__eq__(other)
+ID = "ID"
 
 class GeppettoDataSourceException(Exception): pass
 
@@ -44,7 +35,6 @@ class DataSourceService(metaclass=ServiceCreator):
     def __init__(self, configuration: DataSource, model_access: GeppettoModelAccess):
         self.configuration = configuration
         self.model_access = model_access
-        self.ID = "ID"
 
     def fetch_variable(self):
         raise NotImplemented
@@ -81,10 +71,8 @@ class DataSourceService(metaclass=ServiceCreator):
         final_results = QueryResults(header=next(iter(results.keys())).header)
         first = True
 
-        id_index = final_results.header.index(self.ID)
-        setattr(AQueryResult, '_id_pos', id_index)
-        setattr(AQueryResult, '__eq__', modified_eq)
-        setattr(AQueryResult, '__hash__', modified_hash)
+        id_index = final_results.header.index(ID)
+        set_custom_query_result_hash(id_index)
         
         for result, operator in results.items():
             if final_results.header != result.header:  # TODO test it: may not be supported
@@ -93,20 +81,22 @@ class DataSourceService(metaclass=ServiceCreator):
                 )
 
             if first or operator == BooleanOperator.OR:
-                method = "update"
+                method = final_results.results.update
             
             elif operator == BooleanOperator.AND:
-                method = "intersection_update"
+                method = final_results.results.intersection_update
 
             elif operator == BooleanOperator.NAND:
-                method = "difference_update"
+                method = final_results.results.difference_update
+            else:
+                raise GeppettoDataSourceException(f"Missing operator for query result {result}")
             
-            getattr(final_results.results, method)([r for r in result.results])
+            method(r for r in result.results)
 
             first = False
             
-        setattr(AQueryResult, '__eq__', regular_eq)
-        setattr(AQueryResult, '__hash__', regular_hash)
+        unset_custom_query_result_hash()
+
         return final_results
 
     def process_response(self, response_dict):
